refactor: split Zustand store into slices
